Image (#41)
* Add density attribute to image in index.astro

The density attribute has been added to the 'Mael' image on the index.astro page. This enhancement allows the image to scale correctly on screens with different pixel densities, improving user experience across multiple devices.

* Update dependencies in pnpm-lock.yaml

This commit updates several package versions in pnpm-lock.yaml. Major updates include upgrading the Astro version from 3.5.5 to 3.6.0, the Typescript for markdown package version (^13.0.6 to ^13.0.7), and the esbuild versions. Also, the integrity checksums for the majority of packages have been updated.